move randomString to Executable
