+++ ganga/GangaCore/Lib/Executable/__init__.py
@@ -1,2 +1,2 @@
 
-from .Executable import Executable, RTHandler, LCGRTHandler, gLiteRTHandler
+from .Executable import Executable, RTHandler, LCGRTHandler, gLiteRTHandler, randomString
